Moved all uses of cereal to headers with the suffix '_io.hpp'.
