+++ src/openMVG/cameras/Camera_Pinhole.hpp
@@ -289,13 +289,7 @@ class Pinhole_Intrinsic : public IntrinsicBase
     * @param ar Archive
     */
     template <class Archive>
-    void save( Archive & ar ) const
-    {
-      IntrinsicBase::save(ar);
-      ar( cereal::make_nvp( "focal_length", K_( 0, 0 ) ) );
-      const std::vector<double> pp = {K_( 0, 2 ), K_( 1, 2 )};
-      ar( cereal::make_nvp( "principal_point", pp ) );
-    }
+    inline void save( Archive & ar ) const;
 
 
     /**
@@ -303,15 +297,7 @@ class Pinhole_Intrinsic : public IntrinsicBase
     * @param ar Archive
     */
     template <class Archive>
-    void load( Archive & ar )
-    {
-      IntrinsicBase::load(ar);
-      double focal_length;
-      ar( cereal::make_nvp( "focal_length", focal_length ) );
-      std::vector<double> pp( 2 );
-      ar( cereal::make_nvp( "principal_point", pp ) );
-      *this = Pinhole_Intrinsic( w_, h_, focal_length, pp[0], pp[1] );
-    }
+    inline void load( Archive & ar );
 
     /**
     * @brief Clone the object
@@ -325,9 +311,5 @@ class Pinhole_Intrinsic : public IntrinsicBase
 
 } // namespace cameras
 } // namespace openMVG
-
-#include <cereal/types/polymorphic.hpp>
-CEREAL_REGISTER_TYPE_WITH_NAME(openMVG::cameras::Pinhole_Intrinsic, "pinhole");
-CEREAL_REGISTER_POLYMORPHIC_RELATION(openMVG::cameras::IntrinsicBase, openMVG::cameras::Pinhole_Intrinsic);
 
 #endif // #ifndef OPENMVG_CAMERAS_CAMERA_PINHOLE_HPP
