+++ src/openMVG/cameras/Camera_Pinhole_Brown.hpp
@@ -202,22 +202,14 @@ class Pinhole_Intrinsic_Brown_T2 : public Pinhole_Intrinsic
     * @param ar Archive
     */
     template <class Archive>
-    void save( Archive & ar ) const
-    {
-      ar(cereal::base_class<Pinhole_Intrinsic>(this));
-      ar( cereal::make_nvp( "disto_t2", params_ ) );
-    }
+    inline void save( Archive & ar ) const;
 
     /**
     * @brief  Serialization in
     * @param ar Archive
     */
     template <class Archive>
-    void load( Archive & ar )
-    {
-      ar(cereal::base_class<Pinhole_Intrinsic>(this));
-      ar( cereal::make_nvp( "disto_t2", params_ ) );
-    }
+    inline void load( Archive & ar );
 
     /**
     * @brief Clone the object
@@ -254,11 +246,5 @@ class Pinhole_Intrinsic_Brown_T2 : public Pinhole_Intrinsic
 
 } // namespace cameras
 } // namespace openMVG
-
-#include <cereal/types/polymorphic.hpp>
-#include <cereal/types/vector.hpp>
-
-CEREAL_REGISTER_TYPE_WITH_NAME( openMVG::cameras::Pinhole_Intrinsic_Brown_T2, "pinhole_brown_t2" );
-CEREAL_REGISTER_POLYMORPHIC_RELATION(openMVG::cameras::IntrinsicBase, openMVG::cameras::Pinhole_Intrinsic_Brown_T2)
 
 #endif // #ifndef OPENMVG_CAMERAS_CAMERA_PINHOLE_BROWN_HPP
