+++ src/openMVG/cameras/Camera_Spherical.hpp
@@ -201,20 +201,14 @@ using  class_type = Intrinsic_Spherical;
   * @param ar Archive
   */
   template <class Archive>
-  void save( Archive & ar ) const
-  {
-    ar(cereal::base_class<IntrinsicBase>(this));
-  }
+  inline void save( Archive & ar ) const;
 
   /**
   * @brief  Serialization in
   * @param ar Archive
   */
   template <class Archive>
-  void load( Archive & ar )
-  {
-    ar(cereal::base_class<IntrinsicBase>(this));
-  }
+  inline void load( Archive & ar );
 
   /**
   * @brief Clone the object
@@ -229,18 +223,5 @@ using  class_type = Intrinsic_Spherical;
 
 } // namespace cameras
 } // namespace openMVG
-
-#include <cereal/types/polymorphic.hpp>
-#include <cereal/types/vector.hpp>
-
-CEREAL_REGISTER_TYPE_WITH_NAME(openMVG::cameras::Intrinsic_Spherical, "spherical");
-
-namespace cereal
-{
-  // This struct specialization will tell cereal which is the right way to serialize the ambiguity
-  template <class Archive> struct specialize<Archive, openMVG::cameras::Intrinsic_Spherical, cereal::specialization::member_load_save> {};
-}
-
-CEREAL_REGISTER_POLYMORPHIC_RELATION(openMVG::cameras::IntrinsicBase, openMVG::cameras::Intrinsic_Spherical);
 
 #endif // #ifndef OPENMVG_CAMERAS_CAMERA_SPHERICAL_HPP
