+++ src/openMVG/cameras/cameras_io.hpp
@@ -1,0 +1,20 @@
+// This file is part of OpenMVG, an Open Multiple View Geometry C++ library.
+
+// Copyright (c) 2015 Pierre Moulon.
+
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+#ifndef OPENMVG_CAMERAS_CAMERAS_IO_HPP
+#define OPENMVG_CAMERAS_CAMERAS_IO_HPP
+
+#include "openMVG/cameras/cameras.hpp"
+#include "openMVG/cameras/Camera_Intrinsics_io.hpp"
+#include "openMVG/cameras/Camera_Pinhole_io.hpp"
+#include "openMVG/cameras/Camera_Pinhole_Radial_io.hpp"
+#include "openMVG/cameras/Camera_Pinhole_Brown_io.hpp"
+#include "openMVG/cameras/Camera_Pinhole_Fisheye_io.hpp"
+#include "openMVG/cameras/Camera_Spherical_io.hpp"
+
+#endif // OPENMVG_CAMERAS_CAMERAS_IO_HPP
