+++ src/openMVG/features/regions.cpp
@@ -6,23 +6,22 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#ifndef OPENMVG_FEATURES_IO_REGIONS_TYPE_HPP
-#define OPENMVG_FEATURES_IO_REGIONS_TYPE_HPP
+// The <cereal/archives> headers are special and must be included first.
+#include <cereal/archives/json.hpp> 
 
 #include <fstream>
 #include <string>
 #include <vector>
 
-#include "openMVG/features/regions_factory.hpp"
 #include "third_party/stlplus3/filesystemSimplified/file_system.hpp"
 
-#include <cereal/archives/json.hpp>
+#include "openMVG/features/regions_factory_io.hpp"
 
 namespace openMVG {
 namespace features {
 
 // Init the regions_type from an image describer file (used for regions loading)
-inline std::unique_ptr<features::Regions> Init_region_type_from_file
+std::unique_ptr<features::Regions> Init_region_type_from_file
 (
   const std::string & sImage_describer_file
 )
@@ -47,5 +46,3 @@ inline std::unique_ptr<features::Regions> Init_region_type_from_file
 
 } // namespace features
 } // namespace openMVG
-
-#endif // OPENMVG_FEATURES_IO_REGIONS_TYPE_HPP
