+++ src/openMVG/features/regions.hpp
@@ -78,6 +78,11 @@ public:
 
 };
 
+std::unique_ptr<features::Regions> Init_region_type_from_file
+(
+  const std::string & sImage_describer_file
+);
+
 } // namespace features
 } // namespace openMVG
