+++ src/openMVG/features/sift/SIFT_Anatomy_Image_Describer.hpp
@@ -60,7 +60,6 @@ Changes are:
 #include "openMVG/features/sift/sift_keypoint.hpp"
 #include "openMVG/features/sift/sift_KeypointExtractor.hpp"
 
-#include <cereal/cereal.hpp>
 
 namespace openMVG {
 namespace features {
@@ -86,16 +85,7 @@ public:
       root_sift_(root_sift) {}
 
     template<class Archive>
-    void serialize( Archive & ar )
-    {
-      ar(
-        cereal::make_nvp("first_octave", first_octave_),
-        cereal::make_nvp("num_octaves",num_octaves_),
-        cereal::make_nvp("num_scales",num_scales_),
-        cereal::make_nvp("edge_threshold",edge_threshold_),
-        cereal::make_nvp("peak_threshold",peak_threshold_),
-        cereal::make_nvp("root_sift",root_sift_));
-    }
+    inline void serialize( Archive & ar );
 
     // Parameters
     int first_octave_;      // Use original image, or perform an upscale if == -1
@@ -216,10 +206,7 @@ public:
   }
 
   template<class Archive>
-  void serialize( Archive & ar )
-  {
-    ar(cereal::make_nvp("params", params_));
-  }
+  inline void serialize( Archive & ar );
 
 private:
   Params params_;
@@ -228,9 +215,5 @@ private:
 } // namespace features
 } // namespace openMVG
 
-#include <cereal/types/polymorphic.hpp>
-#include <cereal/archives/json.hpp>
-CEREAL_REGISTER_TYPE_WITH_NAME(openMVG::features::SIFT_Anatomy_Image_describer, "SIFT_Anatomy_Image_describer");
-CEREAL_REGISTER_POLYMORPHIC_RELATION(openMVG::features::Image_describer, openMVG::features::SIFT_Anatomy_Image_describer)
 
 #endif // OPENMVG_FEATURES_SIFT_SIFT_ANATOMY_IMAGE_DESCRIBER_HPP
