+++ src/software/SfM/main_ComputeFeatures_OpenCV.cpp
@@ -6,10 +6,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+// The <cereal/archives> headers are special and must be included first.
+#include <cereal/archives/json.hpp>
+
 #include "openMVG/features/io_regions_type.hpp"
 #include "openMVG/image/image_io.hpp"
 #include "openMVG/sfm/sfm.hpp"
-#include <cereal/archives/json.hpp>
 #include "openMVG/system/timer.hpp"
 
 #include "third_party/cmdLine/cmdLine.h"
@@ -131,7 +133,6 @@ public:
 };
 #include <cereal/cereal.hpp>
 #include <cereal/types/polymorphic.hpp>
-#include <cereal/archives/json.hpp>
 CEREAL_REGISTER_TYPE_WITH_NAME(AKAZE_OCV_Image_describer, "AKAZE_OCV_Image_describer");
 CEREAL_REGISTER_POLYMORPHIC_RELATION(openMVG::features::Image_describer, AKAZE_OCV_Image_describer)
