+++ src/software/SfM/main_exportKeypoints.cpp
@@ -6,7 +6,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "openMVG/features/io_regions_type.hpp"
 #include "openMVG/features/svg_features.hpp"
 #include "openMVG/image/image_io.hpp"
 #include "openMVG/sfm/pipelines/sfm_features_provider.hpp"
@@ -24,7 +23,6 @@
 #include <string>
 
 using namespace openMVG;
-using namespace openMVG::matching;
 using namespace openMVG::sfm;
 
 int main(int argc, char ** argv)
